plan-execution/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
